check post exists before delete, update errors

+++ delete.php
@@ -1,21 +1,32 @@
 <?php 
 require_once("config.php");
+require_once("functions.php");
 
 if (isset($_GET["id"])) {
   $postId = htmlspecialchars($_GET["id"]);
 
   if(is_numeric($postId)) {
 
-    // Get Post from DB 
-    $sql = "DELETE FROM posts WHERE id =" . $postId;
+    //
+    // Check if posts exists before deleting 
+    //
+    $post = getPost($connection, $postId);
+    if($post){
 
-    if($result = mysqli_query($connection, $sql)) {
-        // Set a param with success on the url and redirect to welcome
-        header("location: welcome.php?deleted");
+      $sql = "DELETE FROM posts WHERE id =" . $postId;
+
+      if($result = mysqli_query($connection, $sql)) {
+          // Set a param with success on the url and redirect to welcome
+          header("location: welcome.php?deleted");
+      } else {
+        header("location: welcome.php?error");
+      }
     } else {
+      header("location: welcome.php?nopost");
     }
+
   } else {
-     echo('please enter valid id');
+    header("location: welcome.php?nopost");
   }
 
 }
